Basic functionnalities for configuration are working.
A configuration file can be created if none exist and can be parsed
back. The configuration file is using the XML syntax.

Some "comfort" functionnalities are lacking: for instance checking
that elements pointed to by the configuration file do exist is not
done. We could also check if the RSA key files are indeed valid RSA
private and public keys.