Added missing files, moved constant to "utils.h".
The configuration subpart of the project needed a couple of files in
order to work: namely utils.h and my_info.h.
The constant BUFF_SIZE previously in configuration.h will be used in
other subparts of the project hence we moved it to utils.h and renamed
it for additionnal clarity.
Some whitespaces and comments were respectively removed and added.

Signed-off-by: Loudet Julien <loudet.julien@gmail.com>