Added aes encryption
The encryption requires the path of a file on the system. We make
a system call to encrypt it and we get from the system call the key
and iv used for the encryption.
These elements are stored in a file descriptor structure.
Decryption is a WIP, the command in the comments is working. We just
have to implement it.

Signed-off-by: Loudet Julien <loudet.julien@gmail.com>